requirements.txt
sentence-transformers
transformers
huggingface-hub
torch
faiss-cpu
llama-cpp-python
numpy
tqdm
fitz